% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_score_filter.R
\name{apply_score_filter}
\alias{apply_score_filter}
\title{Apply a score filter to kNN predictions}
\usage{
apply_score_filter(
  scData,
  assay = "predictions",
  slot = "data",
  score.filter = 0.5
)
}
\arguments{
\item{scData}{A Seurat object containing the predictions.}

\item{assay}{A character string specifying the assay from which to retrieve the predictions.
Defaults to "predictions".}

\item{slot}{A character string specifying the slot from which to retrieve the data within the assay.
Defaults to "data".}

\item{score.filter}{A numeric value indicating the threshold below which cells are labeled as "Unassigned".
Defaults to 0.5. This should be adjusted empirically, as few cell types in the core requires a higher cutoff
than a core with many cell types.}
}
\value{
A character vector of predicted cell types, with "Unassigned" for cells that do not meet
the score threshold.
}
\description{
This function applies a score filter to the predictions in a specified assay of a Seurat object.
It assigns "Unassigned" to any cell where the maximum prediction score is below the specified
threshold and returns the name of the cell type with the highest score for those above the threshold.
}
\examples{
predictions <- apply_score_filter(scData = scData, 
                                   assay = "predictions", 
                                   slot = "data", 
                                   score.filter = 0.5)

}
