% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_core.R
\name{downsample_core}
\alias{downsample_core}
\title{Downsample the identified core cells to a specified cell count per final label}
\usage{
downsample_core(
  scData,
  target_count = NULL,
  use_median = TRUE,
  labels_to_exclude = c("unknown"),
  seed = 1984
)
}
\arguments{
\item{scData}{A Seurat object containing the single-cell data and the core cell types as \code{final_labels}.}

\item{target_count}{An integer specifying the target count for downsampling. It is overruled by use_median = TRUE.}

\item{use_median}{A logical indicating whether to use the median cell count (TRUE) or the target count (FALSE).}

\item{labels_to_exclude}{A character vector specifying the labels to exclude (default is \code{c("unknown")}).}

\item{seed}{An integer value for setting the seed to ensure reproducibility (default is 1984).}
}
\value{
A downsampled Seurat object with the specified cell level and a similar number of cells per final label.
}
\description{
This function downsamples a Seurat object to have a similar number of core cells per final label.
It randomly selects cells up to a user-defined target count or the median cell count if specified.
}
\examples{
downsampled_seurat <- downsample_core(scData, target_count = 100, use_median = FALSE, 
                                      labels_to_exclude = c("unknown"), seed = 1984)

}
