% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin_cluster_scoring.R
\name{odin_cluster_scoring}
\alias{odin_cluster_scoring}
\title{Combine cell number and scODIN score for labeling cells on the cluster level}
\usage{
odin_cluster_scoring(scData, clustering_column = "seurat_clusters")
}
\arguments{
\item{scData}{A Seurat object containing the clustering data and scODIN scores.}

\item{clustering_column}{The name of the column that contains the clustering information (default is "seurat_clusters").}
}
\value{
The Seurat object with an updated \code{odin_classification} column,
containing the highest scoring label for each cluster.
}
\description{
This function calculates a combined score based on cell number and scODIN score per cluster,
and assigns the top scoring label to each cluster in the given Seurat object.
}
\examples{
scData <- odin_cluster_scoring(scData, "seurat_clusters")

}
