% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odin_scoring.R
\name{odin_scoring}
\alias{odin_scoring}
\title{scODIN scoring function}
\usage{
odin_scoring(
  gene_priority_table,
  scData,
  core_cell_cutoff = 5,
  similarity_threshold = 1,
  accepted_doubles_table,
  cell_level = "Top"
)
}
\arguments{
\item{gene_priority_table}{A tibble containing gene priorities for different cell types. Usually provided in an Excel sheet.}

\item{scData}{A Seurat object containing normalized and scaled scRNA-seq gene expression data.}

\item{core_cell_cutoff}{A numeric value indicating the cutoff for what is considered a core cell type. Default is 5, but should be adjusted lower if using less deeply sequenced dataset.}

\item{similarity_threshold}{A numeric value defining the similarity threshold for determining double labels. Default is 1.}

\item{accepted_doubles_table}{A tibble for accepted double labels. Usually provided in an Excel sheet.}

\item{cell_level}{A string telling which level the analysis should be done on (e.g., "Top", "CD4_T").}
}
\value{
A Seurat object with scODIN scores and metadata added.
}
\description{
This function calculates scODIN scores for cell types based on priority genes from a provided gene priority table and a Seurat object with gene expression values.
}
\examples{
\dontrun{
# Example usage:
gene_priority_table <- read_excel("~/Desktop/241007_TableS1_gene_priority_table.xlsx")
scData <- your_seurat_object # Load your Seurat object
result <- scODIN_scoring(gene_priority_table, scData, core_cell_cutoff = 5, similarity_threshold = 1,  accepted_doubles_table)
}

}
