% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_double_labels.R
\name{simplify_double_labels}
\alias{simplify_double_labels}
\title{Simplify final double labels in scRNA-seq data}
\usage{
simplify_double_labels(scData, accepted_doubles_table, cell_level)
}
\arguments{
\item{scData}{A Seurat object that contains \code{final_labels} in its metadata.}

\item{accepted_doubles_table}{A tibble for accepted double labels. Usually coming from an Excel sheet.}

\item{cell_level}{Character string specifying the subset level (e.g., "CD4_T").}
}
\value{
The Seurat object with updated \code{final_labels} in the metadata.
}
\description{
This function simplifies the final double labels in a Seurat object
by replacing accepted double combinations of cell types with a simplified name provided by the user.
}
\examples{
\dontrun{
scData <- simplify_double_labels(scData, accepted_doubles_table, "CD4_T")
}

}
